refactor: extract public form lifecycle helpers
Move submission orchestration, validation focus handling, and hydration detection out of Form.tsx so the component focuses on wiring Formik, hCaptcha, and rendered form content.
