+++ components/clientComponents/forms/Form/useIsHydrated.ts
@@ -1,0 +1,8 @@
+import { useSyncExternalStore } from "react";
+
+const subscribeToHydration = () => () => {};
+const getHydratedSnapshot = () => true;
+const getServerSnapshot = () => false;
+
+export const useIsHydrated = () =>
+  useSyncExternalStore(subscribeToHydration, getHydratedSnapshot, getServerSnapshot);
